Split main file into multiple files
This commit is a first step toward separating the different components
into separate files so that they can be tested independently more easily
and so that the project is easier to navigate. This commit introduces
some new problems. For example, the use of from module import * is not
ideal and should be removed later.